% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_result_methods.R
\name{as_fmrireg_hrfs}
\alias{as_fmrireg_hrfs}
\alias{as_fmrireg_hrfs.mhrf_result}
\alias{as_fmrireg_hrfs.matrix}
\title{Convert estimated HRFs to \pkg{fmrireg} objects}
\usage{
as_fmrireg_hrfs(x, ...)

\method{as_fmrireg_hrfs}{mhrf_result}(x, prefix = "voxel", ...)

\method{as_fmrireg_hrfs}{matrix}(x, TR, prefix = "voxel", ...)
}
\arguments{
\item{x}{An object containing HRF estimates. Currently methods are
implemented for \code{mhrf_result} and matrix inputs.}

\item{...}{Additional arguments passed to methods.}

\item{prefix}{Character prefix for naming the HRF objects.}

\item{TR}{Numeric sampling interval for the HRF time axis.}
}
\value{
A list of \code{fmrireg::HRF} objects.
}
\description{
Creates a list of \code{fmrihrf::empirical_hrf} objects from either an
\code{mhrf_result} object or a raw matrix of HRF shapes.
}
