% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_distance_memory}
\alias{check_distance_memory}
\title{Check Memory Requirements for Distance Matrix}
\usage{
check_distance_memory(n_points, element_size = 8, safety_factor = 0.5)
}
\arguments{
\item{n_points}{Number of points in the distance matrix}

\item{element_size}{Size of each element in bytes (default: 8 for double)}

\item{safety_factor}{Fraction of available memory to use (default: 0.5)}
}
\value{
NULL (invisibly). Throws error if memory would be exceeded.
}
\description{
Checks if creating a distance matrix would exceed available memory.
}
\keyword{internal}
