% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_ram_feasibility}
\alias{check_ram_feasibility}
\title{Check RAM feasibility for trial precomputation}
\usage{
check_ram_feasibility(T_trials, V, ram_limit_GB)
}
\arguments{
\item{T_trials}{Number of trials.}

\item{V}{Number of voxels.}

\item{ram_limit_GB}{Memory limit in gigabytes.}
}
\value{
Logical indicating whether precomputation is feasible.
}
\description{
Estimates expected memory usage for storing trial-by-voxel matrices and
compares it to a user-provided limit.
}
\keyword{internal}
