% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss_interface.R
\name{coef.mhrf_lss_result}
\alias{coef.mhrf_lss_result}
\title{Extract Coefficients from M-HRF-LSS Fit}
\usage{
\method{coef}{mhrf_lss_result}(object, type = c("condition", "trial", "hrf"), ...)
}
\description{
Extract Coefficients from M-HRF-LSS Fit
}
