% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_result_methods.R
\name{coef.mhrf_result}
\alias{coef.mhrf_result}
\title{Extract coefficients from mhrf_result}
\usage{
\method{coef}{mhrf_result}(object, type = "amplitudes", ...)
}
\arguments{
\item{object}{An mhrf_result object}

\item{type}{Type of coefficients: "amplitudes", "trial_amplitudes", or "hrfs"}

\item{...}{Additional arguments (ignored)}
}
\description{
Extract coefficients from mhrf_result
}
