% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallback_memory_safety.R
\name{combine_chunk_results}
\alias{combine_chunk_results}
\title{Combine Results from Chunks}
\usage{
combine_chunk_results(results_list)
}
\arguments{
\item{results_list}{List of results from each chunk}
}
\value{
Combined results
}
\description{
Combines results from chunked processing
}
\keyword{internal}
