% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_utils.R
\name{compute_hrf_reasonableness}
\alias{compute_hrf_reasonableness}
\title{Compute HRF Reasonableness Score}
\usage{
compute_hrf_reasonableness(hrf_vector, TR = 2)
}
\arguments{
\item{hrf_vector}{Single HRF time course}

\item{TR}{Time repetition in seconds}
}
\value{
Reasonableness score between 0 and 1
}
\description{
Computes a score indicating how physiologically reasonable an HRF is
}
\keyword{internal}
