% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss_interface.R
\name{compute_pipeline_diagnostics}
\alias{compute_pipeline_diagnostics}
\title{Compute Pipeline Diagnostics}
\usage{
compute_pipeline_diagnostics(
  Y_data,
  Y_proj,
  H_shapes,
  Beta_condition,
  Xi_smoothed,
  TR_precision = 1
)
}
\arguments{
\item{Y_data}{Original data matrix}

\item{Y_proj}{Confound-projected data matrix}

\item{H_shapes}{Matrix of reconstructed HRFs}

\item{Beta_condition}{Condition-level beta estimates}

\item{Xi_smoothed}{Smoothed manifold coordinates}

\item{TR_precision}{Sampling rate of the HRFs}
}
\description{
Compute Pipeline Diagnostics
}
\keyword{internal}
