% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundness_improvements.R
\name{compute_solution_quality}
\alias{compute_solution_quality}
\title{Compute Solution Quality Metrics}
\usage{
compute_solution_quality(
  Y_data,
  Y_predicted,
  Xi_matrix = NULL,
  Beta_matrix = NULL,
  lambda_smooth = NULL
)
}
\arguments{
\item{Y_data}{Data matrix (n x V)}

\item{Y_predicted}{Predicted data matrix (n x V)}

\item{Xi_matrix}{Manifold coordinates (m x V)}

\item{Beta_matrix}{Amplitude matrix (k x V)}

\item{lambda_smooth}{Smoothing parameter used}
}
\value{
List of quality metrics
}
\description{
Computes quality metrics for the current solution
}
