% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss.R
\name{coordinates.NeuroVec}
\alias{coordinates.NeuroVec}
\title{Extract voxel coordinates from NeuroVec}
\usage{
\method{coordinates}{NeuroVec}(x)
}
\description{
Extract voxel coordinates from NeuroVec
}
\keyword{internal}
