% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss.R
\name{coordinates.NeuroVol}
\alias{coordinates.NeuroVol}
\title{Extract voxel coordinates from NeuroVol}
\usage{
\method{coordinates}{NeuroVol}(x)
}
\description{
Extract voxel coordinates from NeuroVol
}
\keyword{internal}
