% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss.R
\name{coordinates}
\alias{coordinates}
\title{Get coordinate extraction function based on data type}
\usage{
coordinates(x)
}
\description{
Get coordinate extraction function based on data type
}
\keyword{internal}
