% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroimaging_wrappers.R
\name{create_flobs_library}
\alias{create_flobs_library}
\title{Create FLOBS Library}
\usage{
create_flobs_library(TR_precision, hrf_duration)
}
\description{
Generates a small set of FLOBS-like HRF basis functions as a list of
\code{fmrireg::HRF} objects.
}
\keyword{internal}
