% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroimaging_wrappers.R
\name{create_gamma_grid_library}
\alias{create_gamma_grid_library}
\title{Create Gamma Grid Library}
\usage{
create_gamma_grid_library(TR_precision, hrf_duration)
}
\description{
Generates a grid of gamma HRFs as a list of \code{fmrireg::HRF} objects.
}
\keyword{internal}
