% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_hrf_manifold.R
\name{create_gamma_hrf_library}
\alias{create_gamma_hrf_library}
\title{Create gamma HRF library}
\usage{
create_gamma_hrf_library(TR, n_basis = 5)
}
\description{
Create gamma HRF library
}
\keyword{internal}
