% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroimaging_wrappers.R
\name{create_half_cosine_library}
\alias{create_half_cosine_library}
\title{Create Half-Cosine Library}
\usage{
create_half_cosine_library(TR_precision, hrf_duration, n_basis = 20)
}
\description{
Create Half-Cosine Library
}
\keyword{internal}
