% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss_interface.R
\name{create_hrf_manifold}
\alias{create_hrf_manifold}
\title{Create HRF Manifold from Various Sources}
\usage{
create_hrf_manifold(hrf_library, params, TR, verbose = TRUE)
}
\arguments{
\item{hrf_library}{HRF library specification}

\item{params}{Parameters for manifold creation}

\item{TR}{Repetition time}

\item{verbose}{Logical indicating whether to print messages}
}
\value{
Manifold object
}
\description{
Creates HRF manifold compatible with fmrireg
}
