% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroimaging_wrappers.R
\name{create_manifold_hrf_object}
\alias{create_manifold_hrf_object}
\title{Create Manifold HRF Object}
\usage{
create_manifold_hrf_object(B_reconstructor, name, nbasis)
}
\description{
Wraps the manifold reconstructor matrix as a multi-basis
\code{\link[fmrireg]{HRF}} object using \code{fmrireg::empirical_hrf} and
\code{fmrireg::bind_basis}. Each column of \code{B_reconstructor} becomes one
basis function.
}
\keyword{internal}
