% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss_interface.R
\name{create_trial_matrices}
\alias{create_trial_matrices}
\title{Create Trial-wise Design Matrices (DEPRECATED)}
\usage{
create_trial_matrices(event_model, event_table, sframe)
}
\description{
Create Trial-wise Design Matrices (DEPRECATED)
}
\keyword{internal}
