% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss_interface.R
\name{create_trial_matrices_from_events}
\alias{create_trial_matrices_from_events}
\title{Create Trial-wise Design Matrices from Event Table}
\usage{
create_trial_matrices_from_events(event_table, hrf_obj, sframe)
}
\description{
Create Trial-wise Design Matrices from Event Table
}
\keyword{internal}
