% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{.check_system_requirements}
\alias{.check_system_requirements}
\title{Check System Requirements}
\usage{
.check_system_requirements(n_voxels, n_timepoints, preset)
}
\arguments{
\item{n_voxels}{Number of voxels}

\item{n_timepoints}{Number of timepoints}

\item{preset}{Analysis preset}
}
\value{
System check results
}
\description{
Check System Requirements
}
\keyword{internal}
