% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss.R
\name{.create_design_matrices}
\alias{.create_design_matrices}
\title{Create design matrices from events}
\usage{
.create_design_matrices(events, n_timepoints, TR, hrf_length = 25)
}
\description{
Create design matrices from events
}
\keyword{internal}
