% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallback_memory_safety.R, R/utils.R
\name{.get_available_memory_gb}
\alias{.get_available_memory_gb}
\title{Get available system memory in GB (platform agnostic)}
\usage{
.get_available_memory_gb()

.get_available_memory_gb()
}
\value{
Available memory in GB, or NULL if unable to determine

Available memory in GB
}
\description{
Platform-agnostic function to get available memory.
}
\keyword{internal}
