% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_result_methods.R
\name{.plot_amplitude_distribution}
\alias{.plot_amplitude_distribution}
\title{Plot amplitude distribution}
\usage{
.plot_amplitude_distribution(x, conditions = NULL, ...)
}
\description{
Plot amplitude distribution
}
\keyword{internal}
