% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_result_methods.R
\name{.plot_manifold_coords}
\alias{.plot_manifold_coords}
\title{Plot manifold coordinates}
\usage{
.plot_manifold_coords(x, dims = c(1, 2), ...)
}
\description{
Plot manifold coordinates
}
\keyword{internal}
