% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss.R
\name{.prepare_data_inputs}
\alias{.prepare_data_inputs}
\title{Prepare data inputs}
\usage{
.prepare_data_inputs(Y_data, events, TR, voxel_mask, progress)
}
\description{
Prepare data inputs
}
\keyword{internal}
