% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss.R
\name{.run_manifold_construction_pca}
\alias{.run_manifold_construction_pca}
\title{Run manifold construction using PCA fallback}
\usage{
.run_manifold_construction_pca(L_library, params, progress, logger = NULL)
}
\description{
Run manifold construction using PCA fallback
}
\keyword{internal}
