% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss.R
\name{.run_spatial_smoothing}
\alias{.run_spatial_smoothing}
\title{Run spatial smoothing}
\usage{
.run_spatial_smoothing(Xi_matrix, voxel_coords, params, progress)
}
\description{
Run spatial smoothing
}
\keyword{internal}
