% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss.R
\name{.run_trial_estimation}
\alias{.run_trial_estimation}
\title{Run trial estimation}
\usage{
.run_trial_estimation(Y_data, X_trial_list, hrf_shapes, params, progress)
}
\description{
Run trial estimation
}
\keyword{internal}
