% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{.validate_events}
\alias{.validate_events}
\title{Validate Events Data Frame}
\usage{
.validate_events(events, n_timepoints, TR)
}
\arguments{
\item{events}{Events data frame}

\item{n_timepoints}{Number of timepoints in Y_data}

\item{TR}{Repetition time}
}
\value{
Validated events information
}
\description{
Validate Events Data Frame
}
\keyword{internal}
