% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{.validate_parameters}
\alias{.validate_parameters}
\title{Validate Analysis Parameters}
\usage{
.validate_parameters(TR, preset, n_voxels, user_params = list())
}
\arguments{
\item{TR}{Repetition time}

\item{preset}{Parameter preset name}

\item{n_voxels}{Number of voxels}

\item{user_params}{Additional user parameters}
}
\value{
Validated parameters
}
\description{
Validate Analysis Parameters
}
\keyword{internal}
