% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{.validate_voxel_mask}
\alias{.validate_voxel_mask}
\title{Validate Voxel Mask}
\usage{
.validate_voxel_mask(voxel_mask, n_voxels)
}
\arguments{
\item{voxel_mask}{Voxel mask (logical or numeric)}

\item{n_voxels}{Total number of voxels}
}
\value{
Validated mask information
}
\description{
Validate Voxel Mask
}
\keyword{internal}
