% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{.validate_Y_data}
\alias{.validate_Y_data}
\title{Validate Y_data Input}
\usage{
.validate_Y_data(Y_data)
}
\arguments{
\item{Y_data}{Input data matrix or neuroimaging object}
}
\value{
Validated data information
}
\description{
Validate Y_data Input
}
\keyword{internal}
