% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallback_memory_safety.R
\name{estimate_memory_requirements}
\alias{estimate_memory_requirements}
\title{Estimate Memory Requirements}
\usage{
estimate_memory_requirements(
  n_timepoints,
  n_voxels,
  n_conditions,
  n_trials,
  p_hrf,
  m_manifold
)
}
\arguments{
\item{n_timepoints}{Number of timepoints}

\item{n_voxels}{Number of voxels}

\item{n_conditions}{Number of conditions}

\item{n_trials}{Number of trials}

\item{p_hrf}{HRF length in timepoints}

\item{m_manifold}{Manifold dimension}
}
\value{
List with memory estimates in GB
}
\description{
Estimates memory needed for M-HRF-LSS pipeline
}
