% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_simulation.R
\name{evaluate_amplitude_recovery}
\alias{evaluate_amplitude_recovery}
\title{Evaluate Amplitude Recovery}
\usage{
evaluate_amplitude_recovery(estimated_betas, true_betas, type)
}
\description{
Evaluate Amplitude Recovery
}
\keyword{internal}
