% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_simulation.R
\name{evaluate_hrf_recovery}
\alias{evaluate_hrf_recovery}
\title{Evaluate HRF Recovery}
\usage{
evaluate_hrf_recovery(estimated_hrfs, true_hrfs)
}
\description{
Evaluate HRF Recovery
}
\keyword{internal}
