% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_simulation.R
\name{evaluate_pipeline_performance}
\alias{evaluate_pipeline_performance}
\title{Evaluate Pipeline Performance}
\usage{
evaluate_pipeline_performance(
  estimates,
  ground_truth_hrfs,
  ground_truth_amplitudes,
  noise_level
)
}
\description{
Evaluate Pipeline Performance
}
\keyword{internal}
