% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_simulation.R
\name{evaluate_spatial_patterns}
\alias{evaluate_spatial_patterns}
\title{Evaluate Spatial Patterns}
\usage{
evaluate_spatial_patterns(estimated_betas, true_betas)
}
\description{
Evaluate Spatial Patterns
}
\keyword{internal}
