% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss_interface.R
\name{extract_design_info}
\alias{extract_design_info}
\title{Extract Design Information from fmrireg Event Model}
\usage{
extract_design_info(event_model, sframe, raw_hrf)
}
\arguments{
\item{event_model}{fmrireg event model object}

\item{sframe}{Sampling frame describing acquisition timing}

\item{raw_hrf}{HRF object used for raw design matrices}
}
\description{
Extract Design Information from fmrireg Event Model
}
\keyword{internal}
