% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_report.R
\name{extract_hrf_stats}
\alias{extract_hrf_stats}
\title{Extract HRF Statistics}
\usage{
extract_hrf_stats(H_shapes, TR_precision = 0.1)

extract_hrf_stats(H_shapes, TR_precision = 0.1)
}
\arguments{
\item{H_shapes}{Matrix of HRF shapes (p x V)}

\item{TR_precision}{Time resolution of HRF sampling}
}
\value{
Data frame with HRF statistics per voxel
}
\description{
Computes statistics about estimated HRF shapes
}
\keyword{internal}
