% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_fit_core.R
\name{extract_xi_beta_raw_svd_robust}
\alias{extract_xi_beta_raw_svd_robust}
\title{Wrapper for robust SVD extraction}
\usage{
extract_xi_beta_raw_svd_robust(
  Gamma_coeffs_matrix,
  m_manifold_dim,
  k_conditions,
  ...
)
}
\description{
Wrapper for robust SVD extraction
}
