% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_hrf_fit_methods.R
\name{fitted.manifold_hrf_fit}
\alias{fitted.manifold_hrf_fit}
\title{Fitted values from manifold_hrf_fit}
\usage{
\method{fitted}{manifold_hrf_fit}(object, ...)
}
\arguments{
\item{object}{A manifold_hrf_fit object}

\item{...}{Additional arguments (ignored)}
}
\value{
The fitted values
}
\description{
Fitted values from manifold_hrf_fit
}
