% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_result_methods.R
\name{fitted.mhrf_result}
\alias{fitted.mhrf_result}
\title{Extract fitted values from mhrf_result}
\usage{
\method{fitted}{mhrf_result}(object, ...)
}
\arguments{
\item{object}{An mhrf_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Extract fitted values from mhrf_result
}
