% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundness_improvements.R
\name{format_time}
\alias{format_time}
\title{Format Time Helper}
\usage{
format_time(seconds)
}
\arguments{
\item{seconds}{Number of seconds}
}
\value{
Formatted time string
}
\description{
Formats seconds into human-readable time
}
\keyword{internal}
