% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_simulation.R
\name{generate_bold_data}
\alias{generate_bold_data}
\title{Generate BOLD Data}
\usage{
generate_bold_data(
  ground_truth_hrfs,
  ground_truth_amplitudes,
  design_info,
  noise_level,
  TR
)
}
\description{
Generate BOLD Data
}
\keyword{internal}
