% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_simulation.R
\name{generate_experimental_design}
\alias{generate_experimental_design}
\title{Generate Experimental Design}
\usage{
generate_experimental_design(
  n_timepoints,
  n_trials,
  n_conditions,
  TR,
  hrf_length = NULL
)
}
\description{
Generate Experimental Design
}
\keyword{internal}
