% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_simulation.R
\name{generate_ground_truth_amplitudes}
\alias{generate_ground_truth_amplitudes}
\title{Generate Ground Truth Amplitudes}
\usage{
generate_ground_truth_amplitudes(
  n_voxels,
  n_conditions,
  n_trials,
  activation_patterns
)
}
\description{
Generate Ground Truth Amplitudes
}
\keyword{internal}
