% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_simulation.R
\name{generate_ground_truth_hrfs}
\alias{generate_ground_truth_hrfs}
\title{Generate Ground Truth HRFs}
\usage{
generate_ground_truth_hrfs(n_voxels, hrf_variability, TR, manifold_params)
}
\description{
Generate Ground Truth HRFs
}
\keyword{internal}
