% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss_interface.R, R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Returns \code{y} if \code{x} is \code{NULL}, otherwise \code{x}.}
\usage{
a \%||\% b

a \%||\% b
}
\arguments{
\item{a}{value to check}

\item{b}{default value}

\item{x}{Primary value to return if not \code{NULL}.}

\item{y}{Fallback value to return if \code{x} is \code{NULL}.}
}
\value{
\code{x} if not \code{NULL}, otherwise \code{y}.

a if not NULL else b
}
\description{
Returns left hand side if not NULL, otherwise right hand side
}
