% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_hrf_fit_class.R
\name{is.manifold_hrf_fit}
\alias{is.manifold_hrf_fit}
\title{Check if object is a manifold_hrf_fit}
\usage{
is.manifold_hrf_fit(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical indicating if x is a manifold_hrf_fit object
}
\description{
Check if object is a manifold_hrf_fit
}
