% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_lss.R
\name{lss_compute_r}
\alias{lss_compute_r}
\title{Internal LSS computation using fmrilss}
\usage{
lss_compute_r(Q_dmat_ran, residual_data)
}
\description{
Internal LSS computation using fmrilss
}
\keyword{internal}
