% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_lss.R
\name{lss_fast}
\alias{lss_fast}
\title{Fast LSS implementation}
\usage{
lss_fast(Y, dmat_base, dmat_ran, dmat_fixed = NULL)
}
\arguments{
\item{Y}{Data matrix}

\item{dmat_base}{Base design matrix}

\item{dmat_ran}{Random effects design}

\item{dmat_fixed}{Fixed effects design}
}
\description{
Wrapper for fmrilss package functionality
}
