% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_hrf_fit_methods.R
\name{plot.manifold_hrf_fit}
\alias{plot.manifold_hrf_fit}
\title{Plot manifold_hrf_fit objects}
\usage{
\method{plot}{manifold_hrf_fit}(
  x,
  type = c("hrf", "manifold", "amplitudes", "fit"),
  voxel = NULL,
  conditions = NULL,
  ...
)
}
\arguments{
\item{x}{A manifold_hrf_fit object}

\item{type}{Type of plot to create}

\item{voxel}{For HRF plots, which voxel(s) to plot (default: 1:5)}

\item{conditions}{For amplitude plots, which conditions to show}

\item{...}{Additional plotting arguments}
}
\description{
Plot manifold_hrf_fit objects
}
