% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss_interface.R
\name{plot.mhrf_lss_result}
\alias{plot.mhrf_lss_result}
\title{Plot M-HRF-LSS Results}
\usage{
\method{plot}{mhrf_lss_result}(x, type = c("hrfs", "manifold", "diagnostics"), voxels = NULL, ...)
}
\description{
Plot M-HRF-LSS Results
}
