% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_result_methods.R
\name{plot.mhrf_result}
\alias{plot.mhrf_result}
\title{Plot method for mhrf_result}
\usage{
\method{plot}{mhrf_result}(x, type = "diagnostic", voxels = NULL, conditions = NULL, ...)
}
\arguments{
\item{x}{An mhrf_result object}

\item{type}{Type of plot: "hrf", "amplitudes", "manifold", "diagnostic"}

\item{voxels}{Which voxels to plot (for "hrf" type)}

\item{conditions}{Which conditions to plot (for "amplitudes" type)}

\item{...}{Additional graphical parameters}
}
\description{
Plot method for mhrf_result
}
