% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_hrf_fit_methods.R
\name{plot_amplitude_distribution}
\alias{plot_amplitude_distribution}
\title{Plot amplitude distributions}
\usage{
plot_amplitude_distribution(x, conditions = NULL, ...)
}
\description{
Plot amplitude distributions
}
\keyword{internal}
