% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_hrf_fit_methods.R
\name{predict.manifold_hrf_fit}
\alias{predict.manifold_hrf_fit}
\title{Predict method for manifold_hrf_fit}
\usage{
\method{predict}{manifold_hrf_fit}(object, newdata = NULL, type = "response", ...)
}
\arguments{
\item{object}{A manifold_hrf_fit object}

\item{newdata}{New event data frame (optional)}

\item{type}{Type of prediction ("response" or "hrf")}

\item{...}{Additional arguments}
}
\value{
Predicted values
}
\description{
Predict method for manifold_hrf_fit
}
