% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_lss_corrected.R
\name{prepare_projection_matrix}
\alias{prepare_projection_matrix}
\title{Prepare Projection Matrix}
\usage{
prepare_projection_matrix(Z_confounds, lambda = 1e-06)
}
\arguments{
\item{Z_confounds}{Confound matrix (n x q)}

\item{lambda}{Ridge regularization parameter}
}
\value{
Projection matrix (n x n)
}
\description{
Creates a projection matrix that removes confound space
}
