% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_control.R
\name{print.manifold_control}
\alias{print.manifold_control}
\title{Print method for manifold_control objects}
\usage{
\method{print}{manifold_control}(x, ...)
}
\arguments{
\item{x}{A manifold_control object}

\item{...}{Ignored}
}
\description{
Print method for manifold_control objects
}
