% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_hrf_fit_methods.R
\name{print.manifold_hrf_fit}
\alias{print.manifold_hrf_fit}
\title{Print method for manifold_hrf_fit objects}
\usage{
\method{print}{manifold_hrf_fit}(x, digits = 3, ...)
}
\arguments{
\item{x}{A manifold_hrf_fit object}

\item{digits}{Number of digits to display}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for manifold_hrf_fit objects
}
