% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroimaging_wrappers.R
\name{print.mhrf_manifold}
\alias{print.mhrf_manifold}
\title{Print method for mhrf_manifold objects}
\usage{
\method{print}{mhrf_manifold}(x, ...)
}
\arguments{
\item{x}{An mhrf_manifold object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for mhrf_manifold objects
}
