% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_report.R
\name{print.mhrf_qc_flags}
\alias{print.mhrf_qc_flags}
\title{Print QC Flags}
\usage{
\method{print}{mhrf_qc_flags}(x, ...)
}
\arguments{
\item{x}{An mhrf_qc_flags object}

\item{...}{Additional arguments (not used)}
}
\description{
Print QC Flags
}
