% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_result_methods.R
\name{print.mhrf_result}
\alias{print.mhrf_result}
\title{S3 Methods for mhrf_result Objects}
\usage{
\method{print}{mhrf_result}(x, ...)
}
\arguments{
\item{x}{An mhrf_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Methods for working with results from the M-HRF-LSS analysis
Print method for mhrf_result
}
