% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_result_methods.R
\name{print.summary.mhrf_result}
\alias{print.summary.mhrf_result}
\title{Print method for summary.mhrf_result}
\usage{
\method{print}{summary.mhrf_result}(x, ...)
}
\arguments{
\item{x}{A summary.mhrf_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for summary.mhrf_result
}
