% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrireg_helpers.R
\name{private_level_token}
\alias{private_level_token}
\title{Create a token for a factor level following fmrireg conventions}
\usage{
private_level_token(factor_name, level)
}
\arguments{
\item{factor_name}{Name of the factor variable}

\item{level}{Level value}
}
\value{
Character string combining the factor name and level
}
\description{
Create a token for a factor level following fmrireg conventions
}
\keyword{internal}
