% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallback_memory_safety.R
\name{process_in_chunks}
\alias{process_in_chunks}
\title{Process Voxels in Chunks}
\usage{
process_in_chunks(process_function, n_voxels, chunk_size = 1000, ...)
}
\arguments{
\item{process_function}{Function that processes a subset of voxels}

\item{n_voxels}{Total number of voxels}

\item{chunk_size}{Number of voxels per chunk}

\item{...}{Additional arguments passed to process_function}
}
\value{
Combined results from all chunks
}
\description{
Wrapper to process voxels in memory-efficient chunks
}
