% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_voxelfit_engine_improved.R,
%   R/voxel_fit_core.R
\name{project_out_confounds_core}
\alias{project_out_confounds_core}
\title{Project out confounds (backward compatible wrapper)}
\usage{
project_out_confounds_core(
  Y_data_matrix,
  X_list_of_matrices,
  Z_confounds_matrix = NULL
)

project_out_confounds_core(
  Y_data_matrix,
  X_list_of_matrices,
  Z_confounds_matrix = NULL
)
}
\arguments{
\item{Y_data_matrix}{n x V matrix of fMRI time series data}

\item{X_list_of_matrices}{List of n x p design matrices for each condition}

\item{Z_confounds_matrix}{Optional n x q matrix of confound regressors}
}
\value{
List with Y_proj_matrix and X_list_proj_matrices
}
\description{
Project out confounds (backward compatible wrapper)

Project out confounds from data and design matrices
}
