% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_lss.R
\name{reconstruct_hrf_shapes_core}
\alias{reconstruct_hrf_shapes_core}
\title{Reconstruct HRF shapes from manifold coordinates}
\usage{
reconstruct_hrf_shapes_core(B_reconstructor_matrix, Xi_manifold_coords_matrix)
}
\arguments{
\item{B_reconstructor_matrix}{p x m manifold basis matrix}

\item{Xi_manifold_coords_matrix}{m x V manifold coordinates}
}
\value{
p x V matrix of reconstructed HRF shapes
}
\description{
Reconstruct HRF shapes from manifold coordinates
}
