% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_hrf_fit_methods.R
\name{residuals.manifold_hrf_fit}
\alias{residuals.manifold_hrf_fit}
\title{Residuals from manifold_hrf_fit}
\usage{
\method{residuals}{manifold_hrf_fit}(object, type = "response", ...)
}
\arguments{
\item{object}{A manifold_hrf_fit object}

\item{type}{Type of residuals (default: "response")}

\item{...}{Additional arguments (ignored)}
}
\value{
The residuals
}
\description{
Residuals from manifold_hrf_fit
}
