% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_lss.R
\name{run_lss_for_voxel}
\alias{run_lss_for_voxel}
\title{Run LSS for Single Voxel with Simple Interface}
\usage{
run_lss_for_voxel(y_voxel, X_trial_list, h_voxel, TR = 2)
}
\arguments{
\item{y_voxel}{Data vector for single voxel (n x 1)}

\item{X_trial_list}{List of trial matrices (n x p each)}

\item{h_voxel}{HRF shape vector (p x 1)}

\item{TR}{Repetition time in seconds (currently unused)}
}
\value{
List with beta_trials vector
}
\description{
Run LSS for Single Voxel with Simple Interface
}
