% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_lss.R
\name{run_lss_woodbury_corrected}
\alias{run_lss_woodbury_corrected}
\alias{run_lsa_woodbury}
\title{Run LSS Woodbury for Single Voxel (Backward Compatibility)}
\usage{
run_lss_woodbury_corrected(
  Y_proj_voxel_vector,
  X_trial_onset_list_of_matrices,
  H_shape_voxel_vector,
  P_confound = NULL,
  Z_confounds = NULL,
  lambda = 0
)

run_lsa_woodbury(
  Y_proj_voxel_vector,
  X_trial_onset_list_of_matrices,
  H_shape_voxel_vector,
  P_confound = NULL,
  Z_confounds = NULL,
  lambda = 0
)
}
\arguments{
\item{Y_proj_voxel_vector}{Projected data vector (n x 1)}

\item{X_trial_onset_list_of_matrices}{List of trial design matrices}

\item{H_shape_voxel_vector}{HRF shape vector (p x 1)}

\item{P_confound}{Projection matrix (n x n) - deprecated parameter}

\item{Z_confounds}{Confound matrix (n x q)}

\item{lambda}{Ridge regularization (unused)}
}
\value{
Vector of trial-wise betas
}
\description{
Run LSS Woodbury for Single Voxel (Backward Compatibility)
}
