% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss_interface.R
\name{run_mhrf_lss_standard}
\alias{run_mhrf_lss_standard}
\title{Run Standard M-HRF-LSS Pipeline}
\usage{
run_mhrf_lss_standard(
  Y_data,
  design_info,
  manifold,
  Z_confounds,
  voxel_coords,
  params,
  outlier_weights,
  estimation,
  progress
)
}
\description{
Run Standard M-HRF-LSS Pipeline
}
\keyword{internal}
