% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_simulation.R
\name{run_pipeline_on_simulated_data}
\alias{run_pipeline_on_simulated_data}
\title{Run Pipeline on Simulated Data}
\usage{
run_pipeline_on_simulated_data(
  bold_data,
  design_info,
  ground_truth_hrfs,
  manifold_params,
  pipeline_params,
  verbose
)
}
\description{
Run Pipeline on Simulated Data
}
\keyword{internal}
