% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_voxelfit_engine_improved.R,
%   R/voxel_fit_core.R
\name{solve_glm_for_gamma_core}
\alias{solve_glm_for_gamma_core}
\title{Solve GLM for gamma (improved wrapper)}
\usage{
solve_glm_for_gamma_core(
  Z_list_of_matrices,
  Y_proj_matrix,
  lambda_gamma,
  orthogonal_approx_flag = FALSE
)

solve_glm_for_gamma_core(
  Z_list_of_matrices,
  Y_proj_matrix,
  lambda_gamma,
  orthogonal_approx_flag = FALSE
)
}
\arguments{
\item{Z_list_of_matrices}{List of n x m transformed design matrices}

\item{Y_proj_matrix}{n x V projected data matrix}

\item{lambda_gamma}{Ridge regularization parameter}

\item{orthogonal_approx_flag}{Whether to use orthogonal approximation}
}
\value{
(k*m) x V matrix of gamma coefficients
}
\description{
Solve GLM for gamma (improved wrapper)

Solve GLM for gamma coefficients
}
