% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_result_methods.R
\name{summary.mhrf_result}
\alias{summary.mhrf_result}
\title{Summary method for mhrf_result}
\usage{
\method{summary}{mhrf_result}(object, ...)
}
\arguments{
\item{object}{An mhrf_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for mhrf_result
}
