% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_voxelfit_engine_improved.R
\name{transform_designs_to_manifold_basis_improved}
\alias{transform_designs_to_manifold_basis_improved}
\title{Transform designs to manifold basis (improved)}
\usage{
transform_designs_to_manifold_basis_improved(X_list, B)
}
\arguments{
\item{X_list}{List of n x p design matrices}

\item{B}{p x m manifold basis matrix}
}
\value{
List of n x m transformed design matrices
}
\description{
Vectorized transformation of design matrices to manifold basis
}
