% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_hrf_manifold.R
\name{transform_to_manifold_hrf_fit}
\alias{transform_to_manifold_hrf_fit}
\title{Transform old result format to manifold_hrf_fit}
\usage{
transform_to_manifold_hrf_fit(
  result_old,
  call,
  control,
  n_timepoints,
  n_voxels,
  n_trials,
  n_conditions,
  TR
)
}
\description{
Transform old result format to manifold_hrf_fit
}
\keyword{internal}
