% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold_hrf_fit_class.R
\name{validate_manifold_hrf_fit}
\alias{validate_manifold_hrf_fit}
\title{Validate manifold_hrf_fit object}
\usage{
validate_manifold_hrf_fit(x)
}
\arguments{
\item{x}{A manifold_hrf_fit object}
}
\value{
TRUE if valid, otherwise throws an error
}
\description{
Internal function to validate the structure of a manifold_hrf_fit object.
}
\keyword{internal}
