% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_voxelfit_engine_improved.R
\name{VoxelWiseGLM}
\alias{VoxelWiseGLM}
\title{VoxelWiseGLM R6 Class for Efficient Voxel-wise Analysis}
\description{
An object-oriented implementation of the voxel-wise GLM fitting engine
with pre-computation of shared matrices and support for parallelization.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VoxelWiseGLM-new}{\code{VoxelWiseGLM$new()}}
\item \href{#method-VoxelWiseGLM-project_out_confounds}{\code{VoxelWiseGLM$project_out_confounds()}}
\item \href{#method-VoxelWiseGLM-fit}{\code{VoxelWiseGLM$fit()}}
\item \href{#method-VoxelWiseGLM-fit_parallel}{\code{VoxelWiseGLM$fit_parallel()}}
\item \href{#method-VoxelWiseGLM-clone}{\code{VoxelWiseGLM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VoxelWiseGLM-new"></a>}}
\if{latex}{\out{\hypertarget{method-VoxelWiseGLM-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VoxelWiseGLM$new(confounds = NULL, ridge_lambda = 1e-06)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{confounds}}{Optional n x q confound matrix}

\item{\code{ridge_lambda}}{Ridge regularization parameter
Project out confounds from data and design matrices}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VoxelWiseGLM-project_out_confounds"></a>}}
\if{latex}{\out{\hypertarget{method-VoxelWiseGLM-project_out_confounds}{}}}
\subsection{Method \code{project_out_confounds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VoxelWiseGLM$project_out_confounds(Y, X_list = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Y}}{n x V data matrix}

\item{\code{X_list}}{List of n x p design matrices}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List with projected Y and X matrices
Fit GLM with QR decomposition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VoxelWiseGLM-fit"></a>}}
\if{latex}{\out{\hypertarget{method-VoxelWiseGLM-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VoxelWiseGLM$fit(Y, X, project_confounds = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Y}}{n x V data matrix}

\item{\code{X}}{n x k design matrix or list of matrices}

\item{\code{project_confounds}}{Whether to project out confounds first}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Coefficient matrix (k x V)
Fit with parallelization over voxels
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VoxelWiseGLM-fit_parallel"></a>}}
\if{latex}{\out{\hypertarget{method-VoxelWiseGLM-fit_parallel}{}}}
\subsection{Method \code{fit_parallel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VoxelWiseGLM$fit_parallel(Y, X, n_cores = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Y}}{n x V data matrix}

\item{\code{X}}{n x k design matrix}

\item{\code{n_cores}}{Number of cores to use}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Coefficient matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VoxelWiseGLM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VoxelWiseGLM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VoxelWiseGLM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
