% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_fit_core.R
\name{apply_intrinsic_identifiability_core}
\alias{apply_intrinsic_identifiability_core}
\title{Apply intrinsic identifiability constraints}
\usage{
apply_intrinsic_identifiability_core(
  Xi_raw_matrix,
  Beta_raw_matrix,
  B_reconstructor_matrix,
  h_ref_shape_vector,
  ident_scale_method = "l2_norm",
  ident_sign_method = "canonical_correlation",
  zero_tol = 1e-08,
  correlation_threshold = 0.001,
  Y_proj_matrix = NULL,
  X_condition_list_proj_matrices = NULL,
  consistency_check = FALSE,
  n_jobs = 1,
  verbose = FALSE
)
}
\description{
Apply intrinsic identifiability constraints
}
