% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss_interface.R
\name{compute_pipeline_diagnostics}
\alias{compute_pipeline_diagnostics}
\title{Compute Pipeline Diagnostics}
\usage{
compute_pipeline_diagnostics(
  Y_data,
  Y_proj,
  H_shapes,
  Beta_condition,
  Xi_smoothed
)
}
\description{
Compute Pipeline Diagnostics
}
\keyword{internal}
