% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss_interface.R
\name{create_hrf_manifold}
\alias{create_hrf_manifold}
\title{Create HRF Manifold from Various Sources}
\usage{
create_hrf_manifold(hrf_library, params, TR, verbose = TRUE)
}
\description{
Internal function to create HRF manifold compatible with fmrireg
}
\keyword{internal}
