% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroimaging_wrappers.R
\name{create_manifold_hrf_object}
\alias{create_manifold_hrf_object}
\title{Create Manifold HRF Object}
\usage{
create_manifold_hrf_object(B_reconstructor, name, nbasis)
}
\description{
Create Manifold HRF Object
}
\keyword{internal}
