% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_utils.R, R/mhrf_lss.R
\name{create_progress_bar}
\alias{create_progress_bar}
\title{Create Progress Bar}
\usage{
create_progress_bar(total)

create_progress_bar(total)
}
\arguments{
\item{total}{Total number of items}

\item{width}{Width of progress bar}
}
\value{
Progress bar object
}
\description{
Creates a simple progress bar for long operations
}
\keyword{internal}
