% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss.R
\name{.compute_qc_metrics}
\alias{.compute_qc_metrics}
\title{Compute QC metrics}
\usage{
.compute_qc_metrics(Y_data, X_condition_list, hrf_shapes, amplitudes,
  manifold_coords, params)
}
\description{
Compute QC metrics
}
\keyword{internal}
