% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhrf_lss_interface.R
\name{extract_design_info}
\alias{extract_design_info}
\title{Extract Design Information from fmrireg Event Model}
\usage{
extract_design_info(event_model, sframe, raw_hrf)
}
\description{
Extract Design Information from fmrireg Event Model
}
\keyword{internal}
