% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_voxelfit_engine_improved.R,
%   R/voxel_fit_core.R
\name{extract_xi_beta_raw_svd_core}
\alias{extract_xi_beta_raw_svd_core}
\title{Extract SVD components (wrapper)}
\usage{
extract_xi_beta_raw_svd_core(
  Gamma_coeffs_matrix,
  m_manifold_dim,
  k_conditions,
  method = c("standard", "robust"),
  ...
)

extract_xi_beta_raw_svd_core(
  Gamma_coeffs_matrix,
  m_manifold_dim,
  k_conditions,
  method = c("standard", "robust"),
  ...
)
}
\arguments{
\item{Gamma_coeffs_matrix}{(km) x V matrix of coefficients}

\item{m_manifold_dim}{Number of manifold dimensions}

\item{k_conditions}{Number of conditions}

\item{method}{Use 'standard' or 'robust' SVD}

\item{...}{Additional arguments passed to the robust implementation}
}
\description{
Extract SVD components (wrapper)

Extract Xi and Beta via SVD
}
