% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{grapes-or-or-grapes}
\alias{grapes-or-or-grapes}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left-hand side value}

\item{y}{Right-hand side value (default)}
}
\value{
\code{x} if not NULL, otherwise \code{y}
}
\description{
Returns the left-hand side if it is not NULL, otherwise returns the right-hand side.
}
\examples{
# Returns 5
NULL \%||\% 5

# Returns 10
10 \%||\% 5

}
