% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R, R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y

x \%||\% y
}
\arguments{
\item{x}{Left-hand side value}

\item{y}{Right-hand side value (default)}

\item{a}{value to check}

\item{b}{default value}
}
\value{
\code{x} if not NULL, otherwise \code{y}

a if not NULL else b
}
\description{
Returns the left-hand side if it is not NULL, otherwise returns the right-hand side.

Returns left hand side if not NULL, otherwise right hand side
}
\examples{
# Returns 5
NULL \%||\% 5

# Returns 10
10 \%||\% 5

}
