% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Default value operator}
\usage{
x \%||\% y
}
\arguments{
\item{a}{value to check}

\item{b}{default value}
}
\value{
a if not NULL else b
}
\description{
Returns left hand side if not NULL, otherwise right hand side
}
