% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_lss.R
\name{lss_fast}
\alias{lss_fast}
\title{Fast LSS implementation using fmrilss}
\usage{
lss_fast(Y, dmat_base, dmat_ran, dmat_fixed = NULL)
}
\description{
Fast LSS implementation using fmrilss
}
\keyword{internal}
