% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qc_summary.R
\name{plot_qc_summary}
\alias{plot_qc_summary}
\title{Plot QC summary for mhrf_result}
\usage{
plot_qc_summary(x)
}
\arguments{
\item{x}{mhrf_result object}
}
\value{
Invisibly returns \code{x}
}
\description{
Plot QC summary for mhrf_result
}
