% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_lss.R
\name{reconstruct_hrf_shapes_core}
\alias{reconstruct_hrf_shapes_core}
\title{Reconstruct HRF Shapes from Manifold Coordinates (Core)}
\usage{
reconstruct_hrf_shapes_core(B_reconstructor_matrix, Xi_smoothed_matrix)
}
\arguments{
\item{B_reconstructor_matrix}{The p x m reconstructor matrix from
\code{get_manifold_basis_reconstructor_core}, where p is HRF length
and m is manifold dimensionality}

\item{Xi_smoothed_matrix}{The m x V matrix of spatially smoothed manifold
coordinates from Component 2, where V is number of voxels}
}
\value{
H_shapes_allvox_matrix A p x V matrix of reconstructed HRF shapes
}
\description{
Transforms smoothed manifold coordinates back to HRF shape space using
the reconstructor matrix from manifold construction.
}
