% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_lss.R
\name{run_lss_for_voxel}
\alias{run_lss_for_voxel}
\title{Run LSS analysis for single voxel (simplified interface)}
\usage{
run_lss_for_voxel(y_voxel, X_trial_list, h_voxel, TR = 2)
}
\arguments{
\item{y_voxel}{Timeseries for single voxel}

\item{X_trial_list}{List of trial onset matrices}

\item{h_voxel}{HRF shape for the voxel}

\item{TR}{Repetition time}
}
\description{
Run LSS analysis for single voxel (simplified interface)
}
