% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_fit_core.R
\name{smart_initialize}
\alias{smart_initialize}
\title{Smart initialization for voxelwise fit}
\usage{
smart_initialize(
  Y_data,
  X_condition_list,
  hrf_canonical,
  use_spatial_clusters = TRUE,
  voxel_coords = NULL,
  m_manifold_dim = 5
)
}
\arguments{
\item{Y_data}{n x V matrix of fMRI data}

\item{X_condition_list}{List of n x p design matrices for each condition}

\item{hrf_canonical}{Canonical HRF shape vector of length p}

\item{use_spatial_clusters}{Whether to use spatial clustering for initialization}

\item{voxel_coords}{V x 3 matrix of voxel coordinates (if use_spatial_clusters = TRUE)}

\item{m_manifold_dim}{Target manifold dimension}
}
\value{
List with initialization values
}
\description{
Smart initialization for voxelwise fit
}
