% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_fit_core.R
\name{smart_initialize}
\alias{smart_initialize}
\title{Smart initialization for voxelwise fit}
\usage{
smart_initialize(
  Y_data,
  X_condition_list,
  hrf_canonical,
  use_spatial_clusters = TRUE,
  voxel_coords = NULL,
  m_manifold_dim = 5
)
}
\description{
Smart initialization for voxelwise fit
}
