% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_voxelfit_engine_improved.R,
%   R/voxel_fit_core.R
\name{solve_glm_for_gamma_core}
\alias{solve_glm_for_gamma_core}
\title{Solve GLM for gamma (improved wrapper)}
\usage{
solve_glm_for_gamma_core(
  Z_list_of_matrices,
  Y_proj_matrix,
  lambda_gamma,
  orthogonal_approx_flag = FALSE
)

solve_glm_for_gamma_core(
  Z_list_of_matrices,
  Y_proj_matrix,
  lambda_gamma,
  orthogonal_approx_flag = FALSE
)
}
\description{
Solve GLM for gamma (improved wrapper)

Solve GLM for gamma coefficients
}
