% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroimaging_wrappers.R
\name{summary.mhrf_manifold}
\alias{summary.mhrf_manifold}
\title{Summary method for mhrf_manifold objects}
\usage{
\method{summary}{mhrf_manifold}(object, ...)
}
\description{
Summary method for mhrf_manifold objects
}
