% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_fit_core.R
\name{transform_designs_to_manifold_basis_core}
\alias{transform_designs_to_manifold_basis_core}
\title{Transform designs to manifold basis}
\usage{
transform_designs_to_manifold_basis_core(
  X_condition_list_proj_matrices,
  B_reconstructor_matrix
)
}
\arguments{
\item{X_condition_list_proj_matrices}{List of n x p projected design matrices}

\item{B_reconstructor_matrix}{p x m manifold basis matrix}
}
\value{
List of n x m transformed design matrices
}
\description{
Transform designs to manifold basis
}
